+++ backend/tests/baserow/contrib/database/airtable/test_airtable_handler.py
@@ -269,6 +269,7 @@ def test_to_baserow_database_export():
             "type": "grid",
             "name": "Grid",
             "order": 1,
+            "row_identifier_type": "id",
             "filter_type": "AND",
             "filters_disabled": False,
             "filters": [],